type fixes for linter

+++ modules/task_3/makarov_a_gauss_filter_col/gauss_filter_col.h
@@ -4,30 +4,30 @@
 
 #include <vector>
 
-std::vector<unsigned short> generate_image(unsigned int w, unsigned int h);
+std::vector<unsigned int> generate_image(unsigned int w, unsigned int h);
 
 std::vector<double> createGaussianKernel(double sigma, unsigned int radius);
 
-std::vector<unsigned short> gaussFilter(
-                               const std::vector<unsigned short>& exp_image,
+std::vector<unsigned int> gaussFilter(
+                               const std::vector<unsigned int>& exp_image,
                                unsigned int w, unsigned int h, double sigma,
                                unsigned int radius);
 
-std::vector<unsigned short> transpose(
-                                   const std::vector<unsigned short>& image,
+std::vector<unsigned int> transpose(
+                                   const std::vector<unsigned int>& image,
                                    unsigned int w, unsigned int h);
 
-std::vector<unsigned short> expand(
-                                   const std::vector<unsigned short>& image,
+std::vector<unsigned int> expand(
+                                   const std::vector<unsigned int>& image,
                                    unsigned int w, unsigned int h,
                                    unsigned int radius);
 
-std::vector<unsigned short> gaussFilterSequential(
-                                   const std::vector<unsigned short>& image,
+std::vector<unsigned int> gaussFilterSequential(
+                                   const std::vector<unsigned int>& image,
                                    unsigned int w, unsigned int h,
                                    double sigma, unsigned int radius);
-std::vector<unsigned short> gaussFilterParallel(
-                                   const std::vector<unsigned short>& image,
+std::vector<unsigned int> gaussFilterParallel(
+                                   const std::vector<unsigned int>& image,
                                    unsigned int w, unsigned int h,
                                    double sigma, unsigned int radius);
